fix advanced copy on mobile
