+++ src/components/Verse/AdvancedCopy/VerseAdvancedCopy.tsx
@@ -1,62 +1,27 @@
 import React, { useState, useEffect, useCallback } from 'react';
 import { useSelector } from 'react-redux';
-import clipboardCopy from 'clipboard-copy';
 import { useRouter } from 'next/router';
 import Verse from 'types/Verse';
 import { selectSelectedTranslations } from 'src/redux/slices/QuranReader/translations';
 import useTranslation from 'next-translate/useTranslation';
-import { getVerseNumberFromKey, generateChapterVersesKeys } from 'src/utils/verse';
-import { getAdvancedCopyRawResult, getAvailableTranslations } from 'src/api';
-import { QuranFont } from 'src/components/QuranReader/types';
+import { generateChapterVersesKeys } from 'src/utils/verse';
+import { getAvailableTranslations } from 'src/api';
 import useSWRImmutable from 'swr/immutable';
 import { makeTranslationsUrl } from 'src/utils/apiPaths';
 import { throwIfError } from 'src/utils/error';
-import Link, { LinkVariant } from 'src/components/dls/Link/Link';
 import { areArraysEqual } from 'src/utils/array';
 import RadioGroup, { RadioGroupOrientation } from '../../dls/Forms/RadioGroup/RadioGroup';
 import Checkbox from '../../dls/Forms/Checkbox/Checkbox';
 import VersesRangeSelector from './VersesRangeSelector';
 import { RangeSelectorType, RangeVerseItem } from './SelectorContainer';
 import styles from './VerseAdvancedCopy.module.scss';
+import { copyText, useTextToCopy } from './advanced-copy-helper';
 
 interface Props {
   verse: Verse;
   children({ onCopy, actionText, ayahSelectionComponent }): React.ReactElement;
 }
 const RESET_BUTTON_TIMEOUT_MS = 5 * 1000;
-
-/**
- * Validate the selected range start and end verse keys. The selection will be invalid in the following cases:
- *
- * 1. One of the two ranges have been cleared and don't have a value.
- * 2. The range start and end verses are the same which is not a valid range. The user should have selected current verse option instead.
- * 3. The range end verse is before the range start verse e.g. from verse 6 -> verse 4.
- *
- * @param {string} selectedRangeStartVerseKey
- * @param {string} selectedRangeEndVerseKey
- * @returns {string|null} if it's null it means the validation passed.
- */
-const validateRangeSelection = (
-  selectedRangeStartVerseKey: string,
-  selectedRangeEndVerseKey: string,
-): string | null => {
-  // if one of them is empty.
-  if (!selectedRangeStartVerseKey || !selectedRangeEndVerseKey) {
-    return 'Range start and end must have a value.';
-  }
-  // if both keys are the same.
-  if (selectedRangeStartVerseKey === selectedRangeEndVerseKey) {
-    return 'Range start and end should be different.';
-  }
-  // if the selected from verse number is higher than the selected to verse number.
-  if (
-    getVerseNumberFromKey(selectedRangeStartVerseKey) >
-    getVerseNumberFromKey(selectedRangeEndVerseKey)
-  ) {
-    return 'The starting verse has to be before the ending verse.';
-  }
-  return null;
-};
 
 const SINGLE_VERSE = 'single';
 const MULTIPLE_VERSES = 'multiple';
@@ -177,56 +142,25 @@ const VerseAdvancedCopy: React.FC<Props> = ({ verse, children }) => {
     }
   };
 
-  const onCopyTextClicked = () => {
-    // if a range is selected, we need to validate it first
-    if (showRangeOfVerses) {
-      const validationError = validateRangeSelection(rangeStartVerse, rangeEndVerse);
-      // if the validation fails
-      if (validationError) {
-        setCustomMessage(validationError);
-        return;
-      }
-    }
-    // by default the from and to will be the current verse.
-    let fromVerse = verse.verseKey;
-    let toVerse = verse.verseKey;
-    // if range of verse was selected
-    if (showRangeOfVerses) {
-      fromVerse = rangeStartVerse;
-      toVerse = rangeEndVerse;
-    }
-    // filter the translations
-    const toBeCopiedTranslations = Object.keys(translations).filter(
-      (translationId) => translations[translationId].shouldBeCopied === true,
-    );
-    getAdvancedCopyRawResult({
-      raw: true,
-      from: fromVerse,
-      to: toVerse,
-      footnote: shouldCopyFootnotes,
-      ...(toBeCopiedTranslations.length > 0 && { translations: toBeCopiedTranslations.join(', ') }), // only include the translations when at least 1 translation has been selected.
-      ...(shouldCopyText && { fields: QuranFont.Uthmani }), // only include the Quranic text if the user chose to.
-    }).then((response) => {
-      // if there is an error
-      if (response.status === 500) {
-        setCustomMessage('Something went wrong, please try again!');
-      } else {
-        clipboardCopy(response.result).then(() => {
-          const objectUrl = window.URL.createObjectURL(
-            new Blob([response.result], { type: 'text/plain' }),
-          );
-          setIsCopied(true);
-          setCustomMessage(
-            <p>
-              Text is copied successfully in your clipboard.{' '}
-              <Link href={objectUrl} download="quran.copy.txt" variant={LinkVariant.Highlight}>
-                Click here
-              </Link>{' '}
-              if you want to download text file.
-            </p>,
-          );
-        });
-      }
+  const textToCopy = useTextToCopy({
+    rangeEndVerse,
+    rangeStartVerse,
+    shouldCopyFootnotes,
+    shouldCopyText,
+    showRangeOfVerses,
+    translations,
+    verseKey: verse.verseKey,
+  });
+
+  const onCopyTextClicked = (e) => {
+    e.preventDefault();
+    copyText({
+      rangeEndVerse,
+      rangeStartVerse,
+      setCustomMessage,
+      setIsCopied,
+      showRangeOfVerses,
+      textToCopy,
     });
   };
